+++ matrix-client/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.0.0
 
 - Expose mkRequest and doRequest
+- Rename DeviceId to DeviceID
 - Return DeviceID out of getTokenOwner
 
 ## 0.1.5.0